Fixed: Can now write with xnvme
The problem was that when the database was closed, the WAL was written
to the file. However only if the WAL has a size greater than 0. Getting
the size relies on the geometry of dev, and since it is not updated on
writes, it is always 0 on new files. Now, we update the geo before
getting the size.
I also add a field for the xnvme file handler, which is used as a
replacement for the 'write' call, which updates the file pointer to the
end of what was just written. This field is can then be used as a kind
of seek mechanism